refactor(tests): remover testes adicionais do ConsentProvider e consolidar em um único arquivo
